backend/requirements-dev.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database (SQLite for development)
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# File Handling & Media
Pillow==10.1.0

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
